pythonback/requirements.txt
Django==3.2.4
django-simpleui==2023.3.1
mysqlclient==2.1.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
pyemail==0.0.1
celery==5.2.7
django-celery-beat
django_apscheduler
django_redis
apscheduler~=3.10.1